% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les-extra.R
\name{find_duplicate}
\alias{find_duplicate}
\title{Find duplicated text}
\usage{
find_duplicate(
  pdfmappe = NULL,
  filnavn = NULL,
  col = FALSE,
  source = TRUE,
  ...
)
}
\arguments{
\item{pdfmappe}{Spesifiserer mappen til pdf filer}

\item{filnavn}{Navn til filen hvis ikke alle filen skal leses}

\item{col}{Kombinere 2 kolonne til en}

\item{source}{Create column geo for source data}

\item{...}{Additional parameters for lespdf function}
}
\value{
A list containing data and dup (duplicate texts) with line index for reference
}
\description{
Find possible duplicated text for control
}
